Use subset of M tokens used as headers in corpus
In the Swedish model, the first token is always L, the second M. The M tokens used here are a subset of all the M tokens in the vocab. So set our header tokens to these, which we already have as the keys of the l_freqs. This might be a special case.