Explicit subset of M, K tokens for drop-down
The swedish corpus has many meter and key tokens that appear only in the body. Get a list of these manually by doing a thousand runs. l_for_m previously got the m list as a by-product, but as need to do this explicitly for k, do that for both.